feat: 認証済みの場合は支払い情報一覧へ遷移し、未認証の場合はトップへ戻す処理を追加した
